Split jobs_storage.py file
No code changes here, all code was just moved
